[Communication] Initial Draft of Battery Module Code
Created an initial draft of the battery module code based on the code
provided by UTHT. Also included the original code for reference. The
new code is half the size of the original and likely burns less power.

This code has not been tested on a real Arduino Micro yet.