[Communication] Initial Draft of LV Power System MCU Code
Very early draft of the LV Power System MCU code. Does not decode
sensor values properly at the moment, as not enough information is
known about the sensors at the moment.

I have a feeling we will have an issue with the I2C communication
with the Arduino Micros. Will need to test them on real hardware.

Have not tested this code on real hardware yet.