[Communication] Updated the Battery Module Code to Use a Timer
Tested the Arduino Micro code on a real Arduino Micro; decided
to use timer interrupts to precalculate the max battery temp.
This was done because we do not want the Due to be waiting too
long (found the code takes about 1030 µs to compute) and wanted
to be able to set the refresh rate of the Micro directly (as
the batteries may rise in temp slower than 15ms.