[Communication] Simple CANBUS Testing
After fabricating the CANBUS, performed simple testing by sending
large files of data from one Arduino Due to another over the
CANBUS. Was able to send up to 1Mb of data over the CANBUS without
error.

Will need a better setup to send larger files as the current method
stores the data in RAM which cannot handle larger files. Will need
to send the data directly to a file.

Ideally, this test should be using the USB to CAN interface the
final design will use. Will wait for this to be acquired.